+++ src/main/java/com/demo/juc/CASDemo.java
@@ -1,4 +1,4 @@
-package com.demo;
+package com.demo.juc;
 
 import java.util.concurrent.atomic.AtomicInteger;
