+++ src/main/java/com/demo/juc/CallableDemo.java
@@ -1,17 +1,8 @@
-package com.demo;
+package com.demo.juc;
 
 import java.util.concurrent.Callable;
 import java.util.concurrent.ExecutionException;
 import java.util.concurrent.FutureTask;
-
-class MyThread implements Callable<Integer>{
-
-    @Override
-    public Integer call() throws Exception {
-        System.out.println("init my thread");
-        return 1024;
-    }
-}
 
 public class CallableDemo {
     public static void main(String[] args) throws ExecutionException, InterruptedException {
@@ -19,5 +10,13 @@ public class CallableDemo {
         new Thread(futureTask, "A").start();
         Integer result = futureTask.get();
         System.out.println("return value: " + result);
+    }
+
+    private static class MyThread implements Callable<Integer>{
+        @Override
+        public Integer call() throws Exception {
+            System.out.println("init my thread");
+            return 1024;
+        }
     }
 }
