+++ src/main/java/com/demo/juc/SynchronousQueueDemo.java
@@ -1,6 +1,5 @@
-package com.demo;
+package com.demo.juc;
 
-import java.util.concurrent.ArrayBlockingQueue;
 import java.util.concurrent.BlockingQueue;
 import java.util.concurrent.SynchronousQueue;
 import java.util.concurrent.TimeUnit;
